backend/requirements/common.txt
django~=3.2.13
pillow
wheel
django-allauth
djangorestframework
djangorestframework-jwt
django-cors-headers~=3.13.0